Add files via upload
These will probably have to be tweaked based on what database we're using, but here's a first crack at a script to create the table and one to summarize it by patient.  For most fields, I kept the max and min of the values, in case there are duplicates within each patient and we want to keep the range.  As we learn more about the fields and we know whether high or low is the bad one for mortality/readmission, we can get rid of a bunch of these fields that wouldn't make sense to include in our model.  I only kept the max for the binary fields, because I figure we'd only be interested in if this field were ever true for the patient.  We should be able to relatively easily turn the summarizedata.sql script into a create table script to populate a much smaller table with only the features we care about.